LPS-64370 move
