+++ portal-impl/src/META-INF/asset-spring.xml
@@ -7,5 +7,5 @@
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd"
 >
-	<bean id="com.liferay.portlet.asset.util.AssetEntryValidatorRegistry" class="com.liferay.portlet.asset.util.AssetEntryValidatorRegistry" />
+	<bean id="com.liferay.portlet.asset.validator.AssetEntryValidatorRegistry" class="com.liferay.portlet.asset.validator.AssetEntryValidatorRegistry" />
 </beans>
