+++ portal-impl/src/com/liferay/portlet/asset/util/AssetEntryValidator.java
@@ -19,7 +19,7 @@ import com.liferay.portal.kernel.exception.PortalException;
 /**
  * @author Brian Wing Shun Chan
  * @deprecated As of 7.0.0, moved to {@link
- *             com.liferay.asset.kernel.util.AssetEntryValidator}
+ *             com.liferay.asset.kernel.validator.AssetEntryValidator}
  */
 @Deprecated
 public interface AssetEntryValidator {
