+++ portal-impl/src/com/liferay/portlet/asset/util/MinimalAssetEntryValidator.java
@@ -16,7 +16,7 @@ package com.liferay.portlet.asset.util;
 
 /**
  * @author Brian Wing Shun Chan
- * @deprecated As of 7.0.0, see {@link
+ * @deprecated As of 7.0.0, replaced by {@link
  *             com.liferay.asset.tags.validator.AtLeastOneTagAssetEntryValidator}
  */
 @Deprecated
