+++ portal-impl/src/com/liferay/portlet/asset/util/packageinfo
@@ -1,1 +1,1 @@
-version 1.1.0
+version 1.0.2
