+++ portal-kernel/src/com/liferay/asset/kernel/validator/packageinfo
